+++ webknossos/webknossos/cli/export_as_tiff.py
@@ -1,22 +1,29 @@
 """This module takes care of exporting tiff images."""
 
 import logging
+import re
 from argparse import Namespace
 from functools import partial
 from math import ceil
 from multiprocessing import cpu_count
 from pathlib import Path
 from typing import Annotated, Any
+from urllib.parse import urlparse
 
 import numpy as np
 import typer
 from PIL import Image
 from scipy.ndimage import zoom
+from upath import UPath
 
+from ..annotation.annotation import _ANNOTATION_URL_REGEX, Annotation
+from ..client import webknossos_context
+from ..client._resolve_short_link import resolve_short_link
 from ..dataset import Dataset, MagView, View
+from ..dataset.dataset import _DATASET_DEPRECATED_URL_REGEX, _DATASET_URL_REGEX
 from ..dataset.defaults import DEFAULT_CHUNK_SHAPE
 from ..geometry import BoundingBox, Mag, Vec3Int
-from ..utils import get_executor_for_args, wait_and_ensure_success
+from ..utils import get_executor_for_args, is_fs_path, wait_and_ensure_success
 from ._utils import (
     DistributionStrategy,
     Vec2Int,
@@ -173,11 +180,9 @@ def export_tiff_stack(
 def main(
     *,
     source: Annotated[
-        Any,
+        str,
         typer.Argument(
-            help="Path to your raw image data.",
-            show_default=False,
-            parser=parse_path,
+            help="Path or URL to your raw image data.",
         ),
     ],
     target: Annotated[
@@ -258,10 +263,47 @@ def main(
             rich_help_panel="Executor options",
         ),
     ] = None,
+    token: Annotated[
+        str | None,
+        typer.Option(
+            help="Authentication token for WEBKNOSSOS instance "
+            "(https://webknossos.org/auth/token).",
+            rich_help_panel="WEBKNOSSOS context",
+            envvar="WK_TOKEN",
+        ),
+    ] = None,
 ) -> None:
     """Export your WEBKNOSSOS dataset to TIFF image data."""
 
-    mag_view = Dataset.open(source).get_layer(layer_name).get_mag(mag)
+    mag_view = None
+    source_path = UPath(source)
+    if not is_fs_path(source_path):
+        url = resolve_short_link(source)
+        parsed = urlparse(url)
+        domain = f"{parsed.scheme}://{parsed.netloc}"
+
+        with webknossos_context(url=domain, token=token):
+            if re.match(_DATASET_URL_REGEX, url) or re.match(
+                _DATASET_DEPRECATED_URL_REGEX, url
+            ):
+                mag_view = Dataset.open_remote(url).get_layer(layer_name).get_mag(mag)
+            elif re.match(_ANNOTATION_URL_REGEX, url):
+                mag_view = (
+                    Annotation.open_as_remote_dataset(annotation_id_or_url=url)
+                    .get_layer(layer_name)
+                    .get_mag(mag)
+                )
+            else:
+                raise ValueError(
+                    "The provided URL does not lead to a dataset or annotation."
+                )
+    else:
+        mag_view = Dataset.open(source_path).get_layer(layer_name).get_mag(mag)
+
+    if mag_view is None:
+        raise ValueError(
+            f"The provided source does not lead to a dataset or annotation. Got: {source}"
+        )
 
     bbox = BoundingBox.from_ndbbox(mag_view.bounding_box) if bbox is None else bbox
     bbox = bbox.align_with_mag(mag_view.mag)