+++ webknossos/webknossos/dataset/dataset.py
@@ -5,7 +5,6 @@ import logging
 import re
 import warnings
 from collections.abc import Callable, Iterable, Mapping, Sequence
-from contextlib import AbstractContextManager
 from datetime import datetime
 from enum import Enum, unique
 from itertools import product
@@ -64,8 +63,10 @@ if TYPE_CHECKING:
 
     from ..administration.user import Team
     from ..client._upload_dataset import LayerToLink
+    from ..client.context import webknossos_context
 
 from ..utils import (
+    cheap_resolve,
     copytree,
     count_defined_values,
     dump_path,
@@ -73,7 +74,6 @@ from ..utils import (
     infer_metadata_type,
     is_fs_path,
     named_partial,
-    resolve_if_fs_path,
     rmtree,
     strip_trailing_slash,
     wait_and_ensure_success,
@@ -323,7 +323,7 @@ class Dataset:
 
         self._read_only = read_only
         self.path: UPath = path
-        self._resolved_path: UPath = resolve_if_fs_path(path)
+        self._resolved_path: UPath = cheap_resolve(path)
 
         if count_defined_values((voxel_size, voxel_size_with_unit)) > 1:
             raise ValueError(
@@ -456,7 +456,7 @@ class Dataset:
         dataset = cls.__new__(cls)
         dataset.path = dataset_path
         dataset._read_only = read_only
-        dataset._resolved_path = resolve_if_fs_path(dataset_path)
+        dataset._resolved_path = cheap_resolve(dataset_path)
         return dataset._init_from_properties(dataset_properties)
 
     @property
@@ -522,15 +522,18 @@ class Dataset:
             datastore_url = _cached_get_upload_datastore(context)
         datastore_api = context.get_datastore_api_client(datastore_url)
         response = datastore_api.dataset_reserve_manual_upload(
-            dataset_announce, token=token
+            dataset_announce=dataset_announce, token=token
         )
         return response.new_dataset_id, response.directory_name
 
     @classmethod
     def trigger_reload_in_datastore(
         cls,
-        dataset_name: str,
-        organization: str,
+        dataset_name_or_url: str | None = None,
+        organization_id: str | None = None,
+        webknossos_url: str | None = None,
+        dataset_id: str | None = None,
+        organization: str | None = None,
         token: str | None = None,
         datastore_url: str | None = None,
     ) -> None:
@@ -543,8 +546,11 @@ class Dataset:
         Cannot be used for local datasets.
 
         Args:
-            dataset_name: Name of dataset to reload
-            organization: Organization ID where dataset is located
+            dataset_name_or_url: Name or URL of dataset to reload
+            dataset_id: ID of dataset to reload
+            organization_id: Organization ID where dataset is located
+            datastore_url: Optional URL to the datastore
+            webknossos_url: Optional URL to the webknossos server
             token: Optional authentication token
 
         Examples:
@@ -560,30 +566,45 @@ class Dataset:
         from ..client._upload_dataset import _cached_get_upload_datastore
         from ..client.context import _get_context
 
-        context = _get_context()
-        upload_url = datastore_url or _cached_get_upload_datastore(context)
-        datastore_api = context.get_datastore_api_client(upload_url)
-        datastore_api.dataset_trigger_reload(organization, dataset_name, token=token)
+        if organization is not None:
+            warn_deprecated("organization", "organization_id")
+            if organization_id is None:
+                organization_id = organization
+            else:
+                raise ValueError(
+                    "Both organization and organization_id were provided. Only one is allowed."
+                )
+
+        (context_manager, dataset_id, _) = cls._parse_remote(
+            dataset_name_or_url,
+            organization_id,
+            None,
+            webknossos_url,
+            dataset_id,
+        )
+
+        with context_manager:
+            context = _get_context()
+            datastore_url = datastore_url or _cached_get_upload_datastore(context)
+            organization_id = organization_id or context.organization_id
+
+            datastore_api = context.get_datastore_api_client(datastore_url)
+            datastore_api.dataset_trigger_reload(
+                organization_id=organization_id, dataset_id=dataset_id, token=token
+            )
 
     @classmethod
     def trigger_dataset_import(
         cls, directory_name: str, organization: str, token: str | None = None
     ) -> None:
-        """Trigger a manual lookup of the dataset in the datastore.
+        """Deprecated. Use `Dataset.trigger_reload_in_datastore` instead."""
+        warn_deprecated("trigger_dataset_import", "trigger_reload_in_datastore")
 
-        Args:
-            directory_name: Directory name of the dataset in the datastore
-            organization: Organization ID where dataset is located
-            token: Optional authentication token
-
-        Examples:
-            ```
-            # Trigger a manual lookup of the dataset in the datastore
-            Dataset.trigger_dataset_upload("l4_sample", "scalable_minds")
-            ```
-        """
-
-        cls.trigger_reload_in_datastore(directory_name, organization, token)
+        cls.trigger_reload_in_datastore(
+            dataset_name_or_url=directory_name,
+            organization_id=organization,
+            token=token,
+        )
 
     @classmethod
     def _disambiguate_remote(
@@ -602,12 +623,12 @@ class Dataset:
         if len(possible_ids) == 0:
             try:
                 dataset_id = current_context.api_client_with_auth.dataset_id_from_name(
-                    dataset_name, organization_id
+                    directory_name=dataset_name, organization_id=organization_id
                 )
                 possible_ids.append(dataset_id)
             except UnexpectedStatusError:
                 raise ValueError(
-                    f"Dataset {dataset_name} not found in organization {organization_id}"
+                    f"Dataset with name {dataset_name} not found in organization {organization_id}"
                 )
         elif len(possible_ids) > 1:
             logger.warning(
@@ -625,7 +646,7 @@ class Dataset:
         sharing_token: str | None = None,
         webknossos_url: str | None = None,
         dataset_id: str | None = None,
-    ) -> tuple[AbstractContextManager, str]:
+    ) -> tuple["webknossos_context", str, str | None]:
         """Parses the given arguments to
         * context_manager that should be entered,
         * dataset_id,
@@ -686,7 +707,7 @@ class Dataset:
         if webknossos_url is None:
             webknossos_url = current_context.url
         webknossos_url = webknossos_url.rstrip("/")
-        context_manager: AbstractContextManager[None] = webknossos_context(
+        context_manager = webknossos_context(
             webknossos_url, token=sharing_token or current_context.token
         )
         if webknossos_url != current_context.url:
@@ -697,10 +718,7 @@ class Dataset:
                     + "Please see https://docs.webknossos.org/api/webknossos/client/context.html to adapt the URL and token."
                 )
                 context_manager = webknossos_context(webknossos_url, None)
-        return (
-            context_manager,
-            dataset_id,
-        )
+        return (context_manager, dataset_id, sharing_token)
 
     @classmethod
     def open_remote(
@@ -738,10 +756,7 @@ class Dataset:
 
         from ..client.context import _get_context
 
-        (
-            context_manager,
-            dataset_id,
-        ) = cls._parse_remote(
+        (context_manager, dataset_id, sharing_token) = cls._parse_remote(
             dataset_name_or_url,
             organization_id,
             sharing_token,
@@ -752,14 +767,15 @@ class Dataset:
         with context_manager:
             wk_context = _get_context()
             token = sharing_token or wk_context.datastore_token
-            api_dataset_info = wk_context.api_client.dataset_info(dataset_id, token)
-            directory_name = api_dataset_info.directory_name
+            api_dataset_info = wk_context.api_client.dataset_info(
+                dataset_id=dataset_id, sharing_token=token
+            )
             organization_id = api_dataset_info.owning_organization
             datastore_url = api_dataset_info.data_store.url
             url_prefix = wk_context.get_datastore_api_client(datastore_url).url_prefix
 
             zarr_path = UPath(
-                f"{url_prefix}/zarr/{organization_id}/{directory_name}/",
+                f"{url_prefix}/zarr/{dataset_id}/",
                 headers={} if token is None else {"X-Auth-Token": token},
                 ssl=SSL_CONTEXT,
             )
@@ -799,10 +815,7 @@ class Dataset:
 
         from ..client._download_dataset import download_dataset
 
-        (
-            context_manager,
-            dataset_id,
-        ) = cls._parse_remote(
+        (context_manager, dataset_id, sharing_token) = cls._parse_remote(
             dataset_name_or_url, organization_id, sharing_token, webknossos_url
         )
 
@@ -1191,7 +1204,7 @@ class Dataset:
         for layer in self.get_segmentation_layers():
             if not layer.attachments.is_empty:
                 raise NotImplementedError(
-                    f"Uploading layers with attachments is not supported yet. Layer {layer.name} has attchments."
+                    f"Uploading layers with attachments is not supported yet. Layer {layer.name} has attachments."
                 )
 
         dataset_id = upload_dataset(
@@ -1626,7 +1639,7 @@ class Dataset:
             except ValueError:
                 continue
             # Mags are only writable if they are local to the dataset
-            resolved_mag_path = resolve_if_fs_path(mag_dir)
+            resolved_mag_path = cheap_resolve(mag_dir)
             read_only = resolved_mag_path.parent != self.resolved_path / layer_name
             layer._add_mag_for_existing_files(
                 mag_dir.name, mag_path=resolved_mag_path, read_only=read_only
@@ -3107,15 +3120,13 @@ class RemoteDataset(Dataset):
         self,
         dataset_path: UPath,
         dataset_id: str,
-        context: AbstractContextManager,
+        context: "webknossos_context",
     ) -> None:
         """Initialize a remote dataset instance.
 
         Args:
             dataset_path: Path to remote dataset location
-            dataset_name: Name of dataset in WEBKNOSSOS
-            organization_id: Organization that owns the dataset
-            sharing_token: Optional token for shared access
+            dataset_id: ID of dataset in WEBKNOSSOS
             context: Context manager for WEBKNOSSOS connection
 
         Raises:
@@ -3193,7 +3204,7 @@ class RemoteDataset(Dataset):
 
         with self._context:
             client = _get_api_client()
-            return client.dataset_info(self._dataset_id)
+            return client.dataset_info(dataset_id=self._dataset_id)
 
     def _update_dataset_info(
         self,
@@ -3225,7 +3236,7 @@ class RemoteDataset(Dataset):
 
         with self._context:
             client = _get_api_client()
-            client.dataset_update(self._dataset_id, info)
+            client.dataset_update(dataset_id=self._dataset_id, updated_dataset=info)
 
     @property
     def metadata(self) -> DatasetMetadata:
@@ -3410,7 +3421,7 @@ class RemoteDataset(Dataset):
 
         with self._context:
             api_sharing_token = _get_api_client().dataset_sharing_token(
-                self._dataset_id
+                dataset_id=self._dataset_id
             )
             return api_sharing_token.sharing_token
 
@@ -3461,7 +3472,7 @@ class RemoteDataset(Dataset):
 
         with self._context:
             client = _get_api_client()
-            client.dataset_update_teams(self._dataset_id, team_ids)
+            client.dataset_update_teams(dataset_id=self._dataset_id, team_ids=team_ids)
 
     @classmethod
     def explore_and_add_remote(
@@ -3499,9 +3510,11 @@ class RemoteDataset(Dataset):
         dataset = ApiDatasetExploreAndAddRemote(
             UPath(dataset_uri).resolve().as_uri(), dataset_name, folder_path
         )
-        context.api_client_with_auth.dataset_explore_and_add_remote(dataset)
+        context.api_client_with_auth.dataset_explore_and_add_remote(dataset=dataset)
 
-        return cls.open_remote(dataset_name, context.organization_id)
+        return cls.open_remote(
+            dataset_name_or_url=dataset_name, organization_id=context.organization_id
+        )
 
     @property
     def folder(self) -> RemoteFolder:
@@ -3572,16 +3585,14 @@ class RemoteDataset(Dataset):
             )
         file_path: UPath
         datastore = context.get_datastore_api_client(datastore_url=datastore_url)
-        api_dataset = context.api_client.dataset_info(self._dataset_id)
+        api_dataset = context.api_client.dataset_info(dataset_id=self._dataset_id)
         directory_name = api_dataset.directory_name
-        organization_id = api_dataset.owning_organization
         assert layer_name is not None, (
             "When you attempt to download a mesh without a tracing_id, the layer_name must be set."
         )
         mesh_download = datastore.download_mesh(
-            mesh_info,
-            organization_id=organization_id or context.organization_id,
-            directory_name=directory_name,
+            mesh_info=mesh_info,
+            dataset_id=self._dataset_id,
             layer_name=layer_name,
             token=token,
         )