+++ webknossos/Changelog.md
@@ -13,10 +13,13 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.10...HEAD)
 
 ### Breaking Changes
+- All internal API client methods (`webknossos.client`) use keyword-only arguments. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
+- The signature of `Dataset.trigger_reload_on_datastore` changed. `dataset_name` has been renamed to `dataset_name_or_url`. Alternatively, a `dataset_id` can be supplied. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
 
 ### Added
 
 ### Changed
+- The API methods now use `v10` API version of Webknossos. Datasets are now referenced by ID instead of dataset name or directory name. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
 
 ### Fixed