+++ webknossos/Changelog.md
@@ -10,15 +10,40 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective _Breaking Changes_ sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.7...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.10...HEAD)
 
 ### Breaking Changes
+- All internal API client methods (`webknossos.client`) use keyword-only arguments. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
+- The signature of `Dataset.trigger_reload_on_datastore` changed. `dataset_name` has been renamed to `dataset_name_or_url`. Alternatively, a `dataset_id` can be supplied. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
 
 ### Added
 
 ### Changed
+- The API methods now use `v10` API version of Webknossos. Datasets are now referenced by ID instead of dataset name or directory name. [#1363](https://github.com/scalableminds/webknossos-libs/pull/1363)
 
 ### Fixed
+
+
+## [2.4.10](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.10) - 2025-08-25
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.9...v2.4.10)
+
+### Added
+- Added support for remote datasets and annotations to `webknossos export-as-tiff`. [#1356](https://github.com/scalableminds/webknossos-libs/pull/1356)
+
+
+## [2.4.9](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.9) - 2025-08-11
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.8...v2.4.9)
+
+### Fixed
+- Fixed a bug where paths of references mags were wrong, if the target dataset's path was a prefix of the source dataset's path. [#1346](https://github.com/scalableminds/webknossos-libs/pull/1346)
+
+
+
+## [2.4.8](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.8) - 2025-08-06
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.7...v2.4.8)
+
+### Added
+- Added support for reading wkw datasets with the new "mag" key in datasource-properties.json, which replaces "wkwResolutions" for datasets written out by the webknossos server. [#1349](https://github.com/scalableminds/webknossos-libs/pull/1349)
 
 
 ## [2.4.7](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.7) - 2025-07-30